// apps/frontend/src/app/api/models/account-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface AccountDto {
  active?: boolean;
  cashBank?: boolean;
  category?: 'Asset' | 'Equity' | 'Expense' | 'Liability' | 'Revenue';
  code?: string;
  createdAt?: string;
  currency?: string;
  fxDefault?: boolean;
  id?: number;
  level?: number;
  name?: string;
  parent?: AccountDto;
  type?: 'H' | 'D';
  updatedAt?: string;
}