// apps/frontend/src/app/api/models/paging-result-account-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { AccountDto } from '../models/account-dto';
import { SortInfo } from '../models/sort-info';
export interface PagingResultAccountDto {
  data?: Array<AccountDto>;
  end?: number;
  page?: number;
  sort?: Array<SortInfo>;
  start?: number;
  totalDisplayRecords?: number;
  totalPages?: number;
  totalRecords?: number;
}